backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
pydantic[email]
supabase
alembic
python-jose[cryptography]
passlib
bcrypt
python-multipart
httpx
slowapi
